refactor: Remove unused scss file, added major UI updates
