+++ client/src/App.tsx
@@ -5,7 +5,7 @@ import { Route, Routes, useLocation } from "react-router-dom";
 import Footer from "@components/Footer/Footer";
 import Navbar from "@/components/Navbar";
 import Home from "@pages/Home/Home";
-import Onboarding from "@pages/Onboarding/Onboarding";
+import Onboarding from "@/pages/Onboarding";
 import Login from "@pages/Authentication/Login/Login";
 import ForgotPassword from "@pages/Authentication/ForgotPassword/ForgotPassword";
 import ForgotPasswordFinal from "@pages/Authentication/ForgotPassword/ForgotPasswordFinal";
@@ -16,7 +16,7 @@ import VerifiedEmail from "@pages/Authentication/Verification/VerifiedEmail";
 import VerifiedAccount from "@pages/Authentication/Verification/VerifyAccount";
 import RegisterBusiness from "@pages/Authentication/Signup/RegisterBusiness/RegisterBusiness";
 import MyAccount from "@pages/AccountSettings/MyAccount";
-import AboutUs from "@pages/AboutUs/AboutUs";
+import AboutUs from "@/pages/AboutUs";
 import ContactSupport from "@pages/ContactSupport/ContactSupport";
 import BusinessContextProvider from "@context/BusinessCtx";
 import ExploreBusiness from "@pages/ExploreBusiness";
