+++ client/src/components/ErrorComponent.tsx
