--- client/src/components/SelectedPlaceholder/placeholder.scss
@@ -1,63 +1,0 @@
-// Days of operation placeholders
-.days-of-operation-container {
-  width: 100%;
-  height: auto;
-  display: flex;
-  align-items: flex-start;
-  justify-items: flex-start;
-  flex-wrap: wrap;
-  margin-top: 10px;
-  gap: 10px;
-
-  .dop-placeholder {
-    width: auto;
-    height: auto;
-    display: flex;
-    align-items: center;
-    justify-items: flex-start;
-    border: none;
-    outline: none;
-    padding: 7px 7px 7px 10px;
-    border-radius: 30px;
-    gap: 1px;
-    transition: all 0.1s ease-in-out;
-    background-color: #eaeaea;
-
-    span {
-      font-size: 11px;
-      letter-spacing: 2px;
-      line-height: 13px;
-      font-weight: 400;
-      color: #292d32;
-      font-family: "Halvetical", sans-serif;
-      margin-left: 5px;
-    }
-
-    button.close-btn {
-      display: flex;
-      align-items: center;
-      justify-items: flex-start;
-      border: none;
-      outline: none;
-      cursor: pointer;
-      transition: all 0.1s ease-in-out;
-      background: none;
-      padding: 0px;
-      margin: 0px;
-
-      .icon {
-        font-size: 5px;
-        color: #a0aec0;
-        transform: scale(0.6);
-      }
-
-      &:hover {
-        transform: scale(0.9);
-      }
-
-      &:active {
-        transform: scale(0.8);
-      }
-    }
-  }
-}
