+++ client/src/components/icons/index.tsx
@@ -7,8 +7,8 @@ import IconMarkup from "./defaultMarkup";
 export const CtaArrow = (props: SvgIconProps) => (
   <IconMarkup {...props}>
     <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
+      fillRule="evenodd"
+      clipRule="evenodd"
       d="M3.0893 9.81429C2.7565 9.86011 2.5 10.1496 2.5 10.4999C2.5 10.8821 2.80526 11.1918 3.18182 11.1918H15.1666L10.8372 15.5678L10.7711 15.6453C10.5725 15.9157 10.5937 16.3002 10.8352 16.5463C11.101 16.8171 11.5327 16.818 11.7995 16.5483L17.2897 10.9997C17.3225 10.9679 17.3521 10.9328 17.378 10.895C17.5639 10.6248 17.5377 10.2499 17.2994 10.0092L11.7994 4.45161L11.7228 4.38478C11.4555 4.18438 11.0767 4.20754 10.8352 4.45372C10.5695 4.72452 10.5704 5.16262 10.8373 5.43226L15.1677 9.80797H3.18182L3.0893 9.81429Z"
       fill="white"
     />
@@ -78,8 +78,8 @@ export const Cancel = (props: SvgIconProps) => (
 export const ArrowUp = (props: SvgIconProps) => (
   <IconMarkup {...props}>
     <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
+      fillRule="evenodd"
+      clipRule="evenodd"
       d="M7.70734 0.293024C7.51981 0.105553 7.26551 0.000236988 7.00034 0.000236988C6.73518 0.000236988 6.48087 0.105553 6.29334 0.293024L0.636343 5.95002C0.540833 6.04227 0.464651 6.15261 0.412242 6.27462C0.359833 6.39662 0.332246 6.52784 0.331092 6.66062C0.329939 6.7934 0.355241 6.92508 0.405521 7.04798C0.455802 7.17087 0.530055 7.28253 0.623948 7.37642C0.717841 7.47031 0.829492 7.54456 0.952389 7.59485C1.07529 7.64513 1.20696 7.67043 1.33974 7.66927C1.47252 7.66812 1.60374 7.64053 1.72575 7.58812C1.84775 7.53572 1.9581 7.45953 2.05034 7.36402L7.00034 2.41402L11.9503 7.36402C12.1389 7.54618 12.3915 7.64698 12.6537 7.6447C12.9159 7.64242 13.1668 7.53725 13.3522 7.35184C13.5376 7.16643 13.6427 6.91562 13.645 6.65342C13.6473 6.39123 13.5465 6.13863 13.3643 5.95002L7.70734 0.293024Z"
       fill="black"
     />
@@ -181,8 +181,8 @@ export const ChevronDown = (props: SvgIconProps) => (
 export const LocationMarker = (props: SvgIconProps) => (
   <IconMarkup {...props}>
     <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
+      fillRule="evenodd"
+      clipRule="evenodd"
       d="M6.18799 5.375C5.46424 5.375 4.87549 5.96375 4.87549 6.68825C4.87549 7.412 5.46424 8 6.18799 8C6.91174 8 7.50049 7.412 7.50049 6.68825C7.50049 5.96375 6.91174 5.375 6.18799 5.375ZM6.18799 9.125C4.84399 9.125 3.75049 8.03225 3.75049 6.68825C3.75049 5.3435 4.84399 4.25 6.18799 4.25C7.53199 4.25 8.62549 5.3435 8.62549 6.68825C8.62549 8.03225 7.53199 9.125 6.18799 9.125Z"
       fill="#0E2D52"
     />
@@ -196,16 +196,16 @@ export const LocationMarker = (props: SvgIconProps) => (
       height="16"
     >
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M0 0.5H12.3746V15.125H0V0.5Z"
         fill="white"
       />
     </mask>
     <g mask="url(#mask0_1144_3457)">
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M6.18701 1.625C3.39551 1.625 1.12451 3.91775 1.12451 6.73475C1.12451 10.319 5.34251 13.811 6.18701 13.997C7.03151 13.8103 11.2495 10.3182 11.2495 6.73475C11.2495 3.91775 8.97851 1.625 6.18701 1.625ZM6.18701 15.125C4.84151 15.125 -0.000488281 10.961 -0.000488281 6.73475C-0.000488281 3.29675 2.77526 0.5 6.18701 0.5C9.59876 0.5 12.3745 3.29675 12.3745 6.73475C12.3745 10.961 7.53251 15.125 6.18701 15.125Z"
         fill="#0E2D52"
       />
