+++ client/src/main.tsx
@@ -6,7 +6,6 @@ import { BrowserRouter } from "react-router-dom";
 
 import App from "./App.tsx";
 import "@styles/index.css";
-import "@styles/index.scss";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
