+++ client/src/pages/Authentication/ForgotPassword/Email.tsx
@@ -7,7 +7,7 @@ import { generateVerificationEmail } from "@/api/auth";
 import { useNavigate } from "react-router-dom";
 import { useState } from "react";
 import { FlexColCenter, FlexColStart, FlexRowCenter } from "@components/Flex";
-import ErrorComponent from "../ErrorComponent";
+import ErrorComponent from "../../../components/ErrorComponent";
 import toast from "react-hot-toast";
 
 const Email = () => {
