+++ client/src/pages/Onboarding/index.tsx
